Renaming api handlers and moving FileFindHandler into base handlr.
